improve: Qt5/Qt6 compatibility with version-aware deprecation macros
- Add QEFI_DEPRECATED_X macro that uses C++17 [[deprecated]] in Qt6
  when C++17 is available, falls back to Q_DECL_DEPRECATED_X for Qt5
- Create internal helper functions to avoid circular deprecation warnings:
  - qefi_validate_load_option()
  - qefi_internal_description_length()
  - qefi_internal_dp_list_length()
  - qefi_internal_optional_data_length()
- Refactor QEFILoadOption::parse() to use internal helpers
- Update deprecated wrapper functions to call internal helpers
- Fix broken multi-line comparison syntax in device path files
- Ensure zero deprecation warnings in internal implementation

+++ qefidpacpi.cpp
@@ -17,8 +17,7 @@ QEFIDevicePath *qefi_parse_dp_acpi_hid(
     if (length != dp_size || length <= 0) return nullptr;
 
     // Check size
-    if (dp_size < QEFI_DEVICE_PATH_HEADER_SIZE +
-        sizeof(quint32) + sizeof(quint32))
+    if (dp_size < (int)(QEFI_DEVICE_PATH_HEADER_SIZE + sizeof(quint32) + sizeof(quint32)))
         return nullptr;
 
     quint8 *dp_inner_pointer = ((quint8 *)dp) +
@@ -43,8 +42,7 @@ QEFIDevicePath *qefi_parse_dp_acpi_hidex(
     if (length != dp_size || length <= 0) return nullptr;
 
     // Check size
-    if (dp_size < QEFI_DEVICE_PATH_HEADER_SIZE +
-        sizeof(quint32) + sizeof(quint32) + sizeof(quint32))
+    if (dp_size < (int)(QEFI_DEVICE_PATH_HEADER_SIZE + sizeof(quint32) + sizeof(quint32) + sizeof(quint32)))
         return nullptr;
 
     quint8 *dp_inner_pointer = ((quint8 *)dp) +
